build: modularize ocp
