+++ solid/src/main/java/solid/ocp/OCPViolated.puml
@@ -1,0 +1,37 @@
+@startuml
+
+class Customer {
+    +buyLotto(money: Int, lottoSeller: LottoSeller, lottoSellerType: LottoSellerType): List<Lottery>
+}
+
+class LottoSeller {
+    +soldLotto(money: Int, lottoSellerType: LottoSellerType): List<Lottery>
+    -LOTTO_PRICE: Int
+    -DISCOUNT_LOTTO_PRICE: Int
+}
+
+enum LottoSellerType {
+    NORMAL
+    DISCOUNT
+}
+
+class Lottery {
+    +numbers: List<Int>
+    +init - validate
+    -MIN_NUMBER: Int
+    -MAX_NUMBER: Int
+    -NUMBER_COUNT: Int
+    -numberRange: IntRange
+}
+
+class LotteryGenerateStrategy {
+    +lotto(): Lottery
+}
+
+Customer --> LottoSeller : uses
+Customer --> LottoSellerType : determines type
+Customer --> Lottery : returns
+LottoSeller ..> LotteryGenerateStrategy : uses
+LottoSeller ..> Lottery : returns
+LotteryGenerateStrategy --> Lottery : creates
+@enduml
